+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 * [-] 移除已废弃的 `Progress` 组件的 `state` prop，请使用 `status` prop 代替。
 * [-] 移除已废弃的 `Progress` 组件的 `auto-succeed` prop，请使用 `autosucceed` prop 代替。
 * [-] 移除已废弃的 `Schedule` 组件的 `shortcuts-display` 值 `expand`/`collapse`，请使用 `inline`/`popup` 代替。
+* [-] 移除已废弃的 `OptionGroup` 组件的 `position` 值 `popout`，请使用 `popup` 代替。
 
 ### 💡 主要变更